Move single instance restriction logic and methods to Config.java to make accessible to both Unitex and GramLab
