+++ src/main/java/fr/gramlab/frames/GramlabFrame.java
@@ -81,6 +81,7 @@ import fr.gramlab.workspace.WorkspaceTreeModel;
 import fr.gramlab.workspace.WorkspaceTreeNode;
 import fr.umlv.unitex.DropTargetManager;
 import fr.umlv.unitex.common.project.manager.GlobalProjectManager;
+import fr.umlv.unitex.config.Config;
 import fr.umlv.unitex.config.ConfigManager;
 import fr.umlv.unitex.files.FileUtil;
 import fr.umlv.unitex.files.PersonalFileFilter;
@@ -183,6 +184,7 @@ public class GramlabFrame extends JFrame {
 
 			@Override
 			public void windowClosed(WindowEvent e) {
+				Config.closeSocket();
 				System.exit(0);
 			}
 		});
